requirements_alt.txt
requests
json
jq
PyJWT
